cartoonize/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==4.4.2
dill==0.3.5.1
entrypoints==0.4
executing==0.8.3
fastapi==0.75.1
flatbuffers==2.0
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.3.0
gast==0.4.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
ipykernel==6.11.0
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
jsonpickle==2.1.0
jupyter-client==7.2.1
jupyter-core==4.9.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.12.0
moviepy==1.0.3
multidict==6.0.2
nest-asyncio==1.5.5
networkx==2.7.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
proglog==0.1.9
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyDeprecate==0.3.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.6.0
PyWavelets==1.3.0
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2022.3.25
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.7.3
torchsummary==1.5.1
torchvision==0.11.3+cu113
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
webdataset==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
yarl==1.7.2
zipp==3.8.0
